+++ src/js/sequences/components/Add.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 import { getDataType } from "../../references/selectors";
 import { routerLocationHasState } from "../../utils/utils";
-import AddBarcodeSequence from "./Barcode/Add";
+import AddBarcodeSequence from "./Barcode/AddBarcodeSequence";
 import AddGenomeSequence from "./Genome/AddGenomeSequence";
 
 export const AddSequence = ({ dataType, show }) => {
