+++ src/js/sequences/components/Barcode/AddBarcodeSequence.tsx
