+++ src/js/sequences/components/Barcode/EditBarcodeSequence.tsx
